Adding styles for list of added to training books for mobile
