+++ src/components/DataPicker/index.js
@@ -1,0 +1,1 @@
+export { default } from './DataPicker';
