+++ src/components/PeriodSelection/PeriodSelection.js
@@ -1,7 +1,7 @@
 import { Formik, Form, Field, ErrorMessage } from 'formik';
 import * as Yup from 'yup';
 import styled from 'styled-components';
-import FormikControl from './FormikControl';
+import FormikControl from '../FormikControl';
 import 'react-datepicker/dist/react-datepicker.css';
 
 const StyledControlsWrapper = styled.div`
