+++ src/components/SetMultiselectValues/SetMultiselectValues.js
@@ -1,4 +1,4 @@
-import SelectBooks from './SelectBooks';
+import SelectBooks from '../SelectBooks';
 import { Field, ErrorMessage, FieldArray } from 'formik';
 export { useState } from 'react';
