+++ src/components/Timer/Timer.style.js
@@ -1,14 +1,5 @@
 import styled from 'styled-components';
 
-export const StyledTimerContainer = styled.div`
-  margin-left: auto;
-  margin-right: auto;
-  width: 654px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
-
 export const StyledTimerWrapper = styled.div`
   font-size: 20px;
   display: flex;
