+++ src/components/TrainingFrom/TrainingForm.js
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import TrainingDataSelection from './TrainingDataSelection';
+import TrainingDataSelection from '../TrainingDataSelection';
 import { TrainingTitle } from './TrainingTitle';
-import Timer from './Timer';
-import { StyledTimerContainer } from './Timer.style';
+import Timer from '../Timer/Timer';
+import { StyledTimerContainer } from './TrainingForm.style';
 
 const TrainingForm = () => {
   const [startDate, setStartDate] = useState(null);
