+++ src/components/TrainingFrom/TrainingForm.style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const StyledTimerContainer = styled.div`
+  margin-left: auto;
+  margin-right: auto;
+  width: 654px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
+
