+++ src/components/TrainingFrom/TrainingTitle.js
